chore: consolidate containsContentItems API and PR feedback
