+++ src/components/app/data/services/index.js
@@ -5,6 +5,5 @@ export * from './course';
 export * from './enterpriseCustomerUser';
 export * from './utils';
 export * from './academies';
-export * from './program';
 export * from './programs';
 export * from './content';
