+++ src/components/app/routes/createAppRouter.jsx
@@ -87,13 +87,13 @@ export default function createAppRouter(queryClient) {
           >
             <Route
               index
-            lazy={async () => {
-              const { ProgramPage, makeProgramLoader } = await import('../../program');
-              return {
-                Component: ProgramPage,
-                loader: makeProgramLoader(queryClient),
-              };
-            }}
+              lazy={async () => {
+                const { ProgramPage, makeProgramLoader } = await import('../../program');
+                return {
+                  Component: ProgramPage,
+                  loader: makeProgramLoader(queryClient),
+                };
+              }}
             />
             <Route
               path="progress"
